refactor: remove interfaces of single components
