+++ src/main/java/com/programmers/app/App.java
@@ -1,27 +1,25 @@
 package com.programmers.app;
 
-import com.programmers.app.instances.GeneralInstances;
-import com.programmers.app.instances.SelectiveInstances;
+import com.programmers.app.config.InitialConfig;
+import com.programmers.app.config.AppConfig;
 import com.programmers.app.menu.Menu;
 import com.programmers.app.mode.Mode;
 
 public class App {
 
-    private final GeneralInstances generalInstances;
     private final Mode mode;
-    private final SelectiveInstances selectiveInstances;
+    private final AppConfig appConfig;
 
-    public App(GeneralInstances generalInstances) {
-        this.generalInstances = generalInstances;
-        mode = generalInstances.getModeSelector().select();
-        selectiveInstances = new SelectiveInstances(generalInstances, mode);
+    public App(InitialConfig initialConfig) {
+        mode = initialConfig.getModeSelector().select();
+        appConfig = new AppConfig(initialConfig, mode);
     }
 
     public void run() {
         Menu menu;
         while (true) {
-            menu = selectiveInstances.getMenuSelector().select();
-            selectiveInstances.getMenuExecuter().execute(menu);
+            menu = appConfig.getMenuSelector().select();
+            appConfig.getMenuExecuter().execute(menu);
         }
     }
 }
