+++ src/main/java/com/programmers/app/config/AppConfig.java
@@ -1,4 +1,4 @@
-package com.programmers.app.instances;
+package com.programmers.app.config;
 
 import com.programmers.app.book.controller.BookController;
 import com.programmers.app.book.controller.BookControllerImpl;
@@ -7,27 +7,25 @@ import com.programmers.app.book.repository.TestBookRepository;
 import com.programmers.app.book.service.BookService;
 import com.programmers.app.book.service.TestBookService;
 import com.programmers.app.menu.MenuExecuter;
-import com.programmers.app.menu.MenuExecuterImpl;
 import com.programmers.app.menu.MenuSelector;
-import com.programmers.app.menu.MenuSelectorImpl;
 import com.programmers.app.mode.Mode;
 import com.programmers.app.timer.TestTimerManager;
 import com.programmers.app.timer.TimerManger;
 
-public class SelectiveInstances {
+public class AppConfig {
 
-    private final GeneralInstances generalInstances;
+    private final InitialConfig initialConfig;
     private final MenuSelector menuSelector;
     private final MenuExecuter menuExecuter;
 
-    public SelectiveInstances(GeneralInstances generalInstances, Mode mode) {
-        this.generalInstances = generalInstances;
+    public AppConfig(InitialConfig initialConfig, Mode mode) {
+        this.initialConfig = initialConfig;
         this.menuSelector = generateMenuSelector();
         this.menuExecuter = generateMenuExecuter(mode);
     }
 
     private MenuSelector generateMenuSelector() {
-        return new MenuSelectorImpl(generalInstances.getCommunicationAgent());
+        return new MenuSelector(initialConfig.getCommunicationAgent());
     }
 
     public MenuSelector getMenuSelector() {
@@ -44,16 +42,16 @@ public class SelectiveInstances {
         return new TestTimerManager();
     }
 
-    private BookService generateBookSevice(Mode mode) {
+    private BookService generateBookService(Mode mode) {
         return new TestBookService(generateBookRepository(mode), generateTimerManager(mode));
     }
 
     private BookController generateBookController(Mode mode) {
-        return new BookControllerImpl(generateBookSevice(mode), generalInstances.getCommunicationAgent());
+        return new BookControllerImpl(generateBookService(mode), initialConfig.getCommunicationAgent());
     }
 
     private MenuExecuter generateMenuExecuter(Mode mode) {
-        return new MenuExecuterImpl(generateBookController(mode));
+        return new MenuExecuter(generateBookController(mode));
     }
 
     public MenuExecuter getMenuExecuter() {
