+++ src/main/java/com/programmers/app/menu/MenuExecuter.java
@@ -1,6 +1,39 @@
 package com.programmers.app.menu;
 
-public interface MenuExecuter {
+import com.programmers.app.book.controller.BookController;
 
-    void execute(Menu menu);
+public class MenuExecuter {
+    private final BookController bookController;
+
+    public MenuExecuter(BookController bookController) {
+        this.bookController = bookController;
+    }
+
+    public void execute(Menu menu) {
+            switch (menu) {
+                case REGISTER:
+                    bookController.register();
+                    break;
+                case REFERNCEALL:
+                    bookController.findAllBooks();
+                    break;
+                case SEARCHTITLE:
+                    bookController.searchBookByTitle();
+                    break;
+                case BORROW:
+                    bookController.borrowBook();
+                    break;
+                case RETURN:
+                    bookController.returnBook();
+                    break;
+                case REPORTLOST:
+                    bookController.reportLostBook();
+                    break;
+                case DELETE:
+                    bookController.deleteBook();
+                    break;
+                case EXIT:
+                    System.exit(0);
+            }
+    }
 }
